+++ runelite-client/src/main/java/net/runelite/client/plugins/menuentryswapper/util/BurningAmuletMode.java
@@ -23,17 +23,17 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.easyscape.util;
+package net.runelite.client.plugins.menuentryswapper.util;
 
-public enum NecklaceOfPassageMode
+public enum BurningAmuletMode
 {
-	WIZARDS_TOWER("Wizard's Tower"),
-	THE_OUTPOST("The Outpost"),
-	EAGLES_EYRIE("Eagle's Eyrie");
+	CHAOS_TEMPLE("Chaos Temple"),
+	BANDIT_CAMP("Bandit Camp"),
+	LAVA_MAZE("Lava Maze");
 
 	private final String name;
 
-	NecklaceOfPassageMode(String name)
+	BurningAmuletMode(String name)
 	{
 		this.name = name;
 	}
