+++ runelite-client/src/main/java/net/runelite/client/plugins/menuentryswapper/util/CombatBraceletMode.java
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2019, Alan Baumgartner <https://github.com/alanbaumgartner>
+ * Copyright (c) 2018, https://runelitepl.us
+ * Copyright (c) 2018, Kyle <https://github.com/kyleeld>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,19 +23,18 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.easyscape.util;
+package net.runelite.client.plugins.menuentryswapper.util;
 
-public enum GamesNecklaceMode
+public enum CombatBraceletMode
 {
-	BURTHORPE("Burthorpe"),
-	BARBARIAN_OUTPOST("Barbarian Outpost"),
-	CORPOREAL_BEAST("Corporeal Beast"),
-	TEARS_OF_GUTHIX("Tears of Guthix"),
-	WINTERTODT("Wintertodt Camp");
+	WARRIORS_GUILD("Warriors' Guild"),
+	CHAMPIONS_GUILD("Champions' Guild"),
+	EDGEVILLE_MONASTERY("Edgeville Monstery"),
+	RANGING_GUILD("Ranging Guild");
 
 	private final String name;
 
-	GamesNecklaceMode(String name)
+	CombatBraceletMode(String name)
 	{
 		this.name = name;
 	}
