+++ runelite-client/src/main/java/net/runelite/client/plugins/menuentryswapper/util/GloryMode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Snakk <http://github.com/SnakkSnokk>
+ * Copyright (c) 2019, Alan Baumgartner <https://github.com/alanbaumgartner>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,21 +22,21 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package net.runelite.client.plugins.menuentryswapper.util;
 
-package net.runelite.client.plugins.menuentryswapper;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum ObeliskMode
+public enum GloryMode
 {
-	ACTIVATE("Activate"),
-	SET_DESTINATION("Set destination"),
-	TELEPORT_TO_DESTINATION("Teleport to destination");
+	EDGEVILLE("Edgeville"),
+	KARAMJA("Karamja"),
+	DRAYNOR_VILLAGE("Draynor Village"),
+	AL_KHARID("Al Kharid");
 
 	private final String name;
+
+	GloryMode(String name)
+	{
+		this.name = name;
+	}
 
 	@Override
 	public String toString()
