+++ runelite-client/src/main/java/net/runelite/client/plugins/menuentryswapper/util/OccultAltarMode.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 2018, https://runelitepl.us
- * Copyright (c) 2018, Kyle <https://github.com/kyleeld>
+ * Copyright (c) 2019, Spedwards <https://github.com/Spedwards>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -23,23 +22,21 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.easyscape.util;
+package net.runelite.client.plugins.menuentryswapper.util;
 
-public enum SkillsNecklaceMode
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum OccultAltarMode
 {
-	FISHING_GUILD("Fishing Guild"),
-	MINING_GUILD("Mining Guild"),
-	CRAFTING_GUILD("Crafting Guild"),
-	COOKING_GUILD("Cooking Guild"),
-	WOODCUTTING_GUILD("Woodcutting Guild"),
-	FARMING_GUILD("Farming Guild");
+	ANCIENT("Ancient"),
+	LUNAR("Lunar"),
+	ARCEUUS("Arceuus"),
+	VENERATE("Venerate");
 
 	private final String name;
-
-	SkillsNecklaceMode(String name)
-	{
-		this.name = name;
-	}
 
 	@Override
 	public String toString()
