+++ runelite-client/src/main/java/net/runelite/client/plugins/menuentryswapper/util/SkillsNecklaceMode.java
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2018, Snakk <http://github.com/SnakkSnokk>
+ * Copyright (c) 2018, https://runelitepl.us
+ * Copyright (c) 2018, Kyle <https://github.com/kyleeld>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,22 +23,23 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package net.runelite.client.plugins.menuentryswapper.util;
 
-package net.runelite.client.plugins.menuentryswapper;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum HouseMode
+public enum SkillsNecklaceMode
 {
-	ENTER("Enter"),
-	HOME("Home"),
-	BUILD_MODE("Build mode"),
-	FRIENDS_HOUSE("Friend's House");
+	FISHING_GUILD("Fishing Guild"),
+	MINING_GUILD("Mining Guild"),
+	CRAFTING_GUILD("Crafting Guild"),
+	COOKING_GUILD("Cooking Guild"),
+	WOODCUTTING_GUILD("Woodcutting Guild"),
+	FARMING_GUILD("Farming Guild");
 
 	private final String name;
+
+	SkillsNecklaceMode(String name)
+	{
+		this.name = name;
+	}
 
 	@Override
 	public String toString()
